+++ style.css
@@ -1,0 +1,4 @@
+#coach-ai {
+  font-size: 4rem;
+  text-align: center;
+}
